// src/common/enums/appEnum.ts
// 系统级别枚举
export enum LayoutModeEnum {
  Vertical = "vertical",
  Classic = "classic",
  Horizontal = "horizontal",
  Columns = "columns",
  IFrame = "iframe",
  Mixins = "mixins",
}

// 标签栏模式
export enum TabNavModeEnum {
  Simple = "simple",
  Classic = "classic",
  Element = "element",
}

// 系统主题
export enum SystemThemeEnum {
  Dark = "dark",
  Light = "light",
  System = "system",
}

// 菜单主题
export enum MenuThemeEnum {
  Dark = "dark",
  Light = "light",
}

// 标题模式
export enum TitleModeEnum {
  ProjectPage = "projectPage",
  UsernamePage = "usernamePage",
  Project = "project",
  Page = "page",
}

// 语言
export enum LanguageEnum {
  ZhCn = "zh-CN",
  EnUs = "en-US",
}

// 布局大小
export enum LayoutSizeEnum {
  Default = "default",
  Small = "small",
  Large = "large",
}

// 路由页面过渡效果
export enum PageTransitionEnum {
  None = "",
  Fade = "fade",
  SlideLeft = "slide-left",
  SlideTop = "Slide-top",
  SlideBottom = "slide-bottom",
}

// 顶部风格
export enum HeaderStyleEnum {
  Page = "page",
  Bg = "bg",
  Line = "line",
  BgLine = "bgLine",
}